refactor(router): split route to new file, add index-layout route
